+++ customer-service/src/main/java/com/microservico/customer/dto/response/ItemPedidoDtoResponse.java
@@ -1,4 +1,6 @@
 package com.microservico.customer.dto.response;
 
-public record ItemPedidoDtoResponse(Long id, Long produtoId, String nomeProduto, Double precoUnitario, int quantidade) {
+import java.math.BigDecimal;
+
+public record ItemPedidoDtoResponse(Long id, Long produtoId, String nomeProduto, BigDecimal precoUnitario, int quantidade) {
 }
